# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Questerra — player API",
        "version": "v1",
        "description": "Generated by tools/openapi-gen.php from the running router, the controllers'\nvalidate() rules, and a live probe of GET responses. Do not edit by hand: the\nhand-written contract drifted, which is why this file exists.\n\nSemantics a schema cannot carry — that submit stays open after time_expired, that\nfun_game answers score zero, the minimum interval between position posts — live in\nAPK-BUILD-GUIDE.md and still have to be read.",
        "x-behaviour-guides": {
            "API-V1-CONTRACT.md": "8afbdf4e5e3c2c79",
            "APK-BUILD-GUIDE.md": "a83f14c3f7f0267b",
            "APK-WORK-ORDER.md": "927669638c184ad6"
        }
    },
    "servers": [
        {
            "url": "https://questerra-series.com"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Laravel Sanctum personal access token"
            }
        }
    },
    "paths": {
...
        "/api/v1/quiz/attempts": {
            "get": {
                "operationId": "api.v1.quiz.attempts",
                "summary": "api.v1.quiz.attempts",
                "x-auth": "sanctum-bearer",
                "x-access-window": true,
                "x-rate-limit": "api",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "nullable": true,
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100,
                            "x-laravel-rules": "nullable|integer|min:1|max:100"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Observed on a live probe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "attempts": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "completed_at": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    },
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "questionnaire": {
                                                        "type": "object",
                                                        "properties": {
                                                            "id": {
                                                                "description": "nested"
                                                            },
                                                            "title": {
                                                                "description": "nested"
                                                            }
                                                        }
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "total_score": {
                                                        "type": "integer"
                                                    },
                                                    "total_time_seconds": {
                                                        "type": "integer"
                                                    }
                                                }
                                            }
                                        },
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "total_attempts": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}